requirements.txt
pandas
numpy
yfinance
matplotlib
openpyxl
transformer
requests
beautifulsoup4
scikit-learn
